fix: Switching tabs
